refactor: Line Class의 Sections을 일급 컬렉션으로 분리함
